+++ src/main/java/com/fesc/tiendaOnline/model/dto/ProductoResponseDTO.java
@@ -1,5 +1,6 @@
 package com.fesc.tiendaOnline.model.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import lombok.Data;
@@ -10,7 +11,7 @@ public class ProductoResponseDTO {
     private UUID idProducto;
     private String nombreProducto;
     private String descripcionProducto;
-    private Double precioProducto;
+    private BigDecimal precioProducto;
     private Integer stockProducto;
     private String urlImagenProducto;
     private String nombreCategoria;
